Scrape of the hieroglyphicus mockup.
This is a draft version of an Egyptian hieroglyph dictionary,
by Boris Jegorović and Ognjan Eskić. Boris asked for collaborators
on the GlyphStudy Yahoo group, but didn't provide a repository,
so this is a just copy of what's available client-side as a basis
for further work.

In particular, there's no php code; the pages have already been
rendered server-side.

The original was at https://hieroglyphicus.000webhostapp.com/
Copyright 2018 Particle Design.