Add a basic mysql container.
Use the mysql base image from docker-library as the basis for a
database container. Add an init.sql file with a table for basic
dictionary entries. This doesn't represent the full schema from
Boris' document, or even a full design. It's just a hack so I
have something to test against.

Build with `docker build -t hgy-mysql sql`
Start an instance with `docker run -d hgy-mysql`

The root password is randomly generated and available
from `docker logs $container_id | grep 'GENERATED ROOT PASSWORD'`

Link this container to the php container instances so port 3306
is available to them for database queries.